Minor refactor to break up the huge lib.rs
